# backend/src/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Christephanie API",
    "version": "v1"
  },
  "paths": {
...
    "/api/payments/intent": {
      "post": {
        "tags": [
          "Payment"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StripePaymentIntentRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StripePaymentIntentRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StripePaymentIntentRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StripePaymentIntentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "tags": [
          "Payment"
        ],
        "parameters": [
          {
            "name": "paymentIntentId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StripePaymentIntentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/payments/contributions": {
      "get": {
        "tags": [
          "Payment"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContributionDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContributionDto": {
        "type": "object",
        "properties": {
          "paymentIntentId": {
            "type": "string",
            "nullable": true
          },
          "guestId": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "integer",
            "format": "int32"
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "giftCategory": {
            "type": "string",
            "nullable": true
          },
          "giftNotes": {
            "type": "string",
            "nullable": true
          },
          "guestName": {
            "type": "string",
            "nullable": true
          },
          "isAnonymous": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "GiftMetaData": {
        "type": "object",
        "properties": {
          "guestId": {
            "type": "string",
            "nullable": true
          },
          "guestName": {
            "type": "string",
            "nullable": true
          },
          "isAnonymous": {
            "type": "boolean"
          },
          "guestEmail": {
            "type": "string",
            "nullable": true
          },
          "giftCategory": {
            "type": "string",
            "nullable": true
          },
          "giftNotes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PaymentError": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "decline_code": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "StripePaymentIntentRequestDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "integer",
            "format": "int32"
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "giftMetaData": {
            "$ref": "#/components/schemas/GiftMetaData"
          }
        },
        "additionalProperties": false
      },
      "StripePaymentIntentResponseDto": {
        "type": "object",
        "properties": {
          "clientSecret": {
            "type": "string",
            "nullable": true
          },
          "paymentIntentId": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "error": {
            "$ref": "#/components/schemas/PaymentError"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Enter 'Bearer' [space] and then your valid token in the text input below.\n\nExample: \"Bearer eyJhbGciOiJIUz...\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}